Fixes potential crash updating RSTFetchedResultsDataSource when contentView isn’t visible
Previously, we ignored NSFetchedResultsController changes when contentView is not in view hierarchy. However, this can cause the contentView to become out-of-sync with the data source, resulting in a crash when contentView is added back to view hierarchy.

Now, if contentView is not in view hierachy we call -[RSTCellContentView reloadData] instead. This ensures contentView is always in sync with the data source, but contentView won’t actually update the cells until it’s visible again.